+++ src/main/java/calculator/service/InputService.java
@@ -3,6 +3,8 @@ package calculator.service;
 import calculator.enums.ErrorMessages;
 import calculator.view.InputView;
 
+import static calculator.util.Validator.validateCustomDelimiter;
+
 public class InputService {
 
     private final InputView inputView;
@@ -19,23 +21,6 @@ public class InputService {
             return input;
         } catch (Exception e) {
             throw new IllegalArgumentException(ErrorMessages.EMPTY_INPUT.getMessage());
-        }
-    }
-
-    private void validateCustomDelimiter(String input) {
-        //TODO: 커스텀 구분자 검증 로직 구현
-        if (input.startsWith("//")) {
-            int newlineIndex = input.indexOf("\\n");
-
-            if (newlineIndex == -1) {
-                throw new IllegalArgumentException(ErrorMessages.MISSING_NEWLINE.getMessage());
-            }
-
-            String delimiter = input.substring(2, newlineIndex);
-
-            if (delimiter.isEmpty()) {
-                throw new IllegalArgumentException(ErrorMessages.MISSING_CUSTOM_DELIMITER.getMessage());
-            }
         }
     }
 }
